Toy Robot Test

+++ src/ToyRobot/Table/CircularTable.hpp
@@ -6,9 +6,8 @@ namespace Robot
     class CircularTable : public ITable
     {
     public:
-
         CircularTable(int diameter) : m_diameter{diameter} {};
-        bool IsValidPosition(float x, float y) const override;
+        bool IsValidPosition(int x, int y) const override;
         void PrintTable(int x, int y, const std::string &facing) override;
 
     private:
