+++ src/ToyRobot/Table/SquircleTable.hpp
@@ -8,7 +8,7 @@ namespace Robot
     {
     public:
         SquircleTable(int width, int height) : m_width{width}, m_height{height} {};
-        bool IsValidPosition(float x, float y) const override;
+        bool IsValidPosition(int x, int y) const override;
         void PrintTable(int x, int y, const std::string &facing) override;
 
     private:
